backup/back/requirements.txt
# requirements.txt
Django==5.0.8
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-axes==6.3.0
django-extensions==3.2.3
Pillow==10.3.0
reportlab==4.1.0
python-dateutil==2.9.0.post0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
weasyprint==62.1